move tests to src
