+++ src/js/model/primitives/Guide.test.js
@@ -1,22 +1,25 @@
-var Guide = require('../src/js/model/primitives/Guide.js');
-var Scale = require('../src/js/model/primitives/Scale.js');
-var guide, scale;
+/* eslint no-unused-expressions:0 */
+var expect = require('chai').expect;
+
+var Guide = require('./Guide');
+var Scale = require('./Scale');
+var scale, axisGuide, legendGuide;
 
 beforeEach(function() {
   var def = {
-      name: "x",
-      padding: 1,
-      points: true,
-      range: "width",
-      round: true,
-      type: "ordinal"
-    };
+    name: 'x',
+    padding: 1,
+    points: true,
+    range: 'width',
+    round: true,
+    type: 'ordinal'
+  };
   scale = new Scale(def.name, def.type, undefined, def.range);
-  axisGuide = new Guide(1, "x", scale._id);
-  legendGuide = new Guide(2, "x", scale._id);
+  axisGuide = new Guide(1, 'x', scale._id);
+  legendGuide = new Guide(2, 'x', scale._id);
 });
 
-describe('Axis Guide', function(){
+describe('Axis Guide', function() {
   it('Guide type axis initializes with type', function() {
     expect(axisGuide.type).to.equal('x');
   });
@@ -42,13 +45,13 @@ describe('Axis Guide', function(){
 
 });
 
-describe('Legend Guide', function(){
+describe('Legend Guide', function() {
   it('Guide type legend initializes with _type', function() {
     expect(legendGuide._type).to.equal('x');
   });
 
   it('Guide type legend initializes with x property', function() {
-    expect(legendGuide["x"]).to.equal(scale._id);
+    expect(legendGuide.x).to.equal(scale._id);
   });
 
   it('Guide type legend initializes with correct properties', function() {
